Backend/backend_2/requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
Werkzeug==3.1.7
